refractored all the files. added new layouts
moved all files into new packages, added layout to make it so the user can go click on add expense or add income, this takes them to the transaction page. this page does not work as of now as it needs categories. category database is in place as well as the transaction database also in place, need the category database to use the transaction  page as it will not work without